service/src/requirements.txt
graphviz
lark-parser
ipdb
nose
